Parsing logic now completely separated, adder functions are still in the main lib.rs file and will stay
